Refactor to move all signals to a separate package
- separate the module for testing signals.

Closes #61
