+++ comment/tests/test_models.py
@@ -6,7 +6,7 @@ from django.db import IntegrityError
 from django.utils import timezone
 
 from comment.conf import settings
-from comment.models import Comment, Flag, FlagInstance, Reaction, ReactionInstance
+from comment.models import Comment, FlagInstance, ReactionInstance
 from comment.tests.base import BaseCommentManagerTest, BaseCommentTest, BaseCommentFlagTest
 
 
@@ -50,21 +50,6 @@ class CommentModelTest(BaseCommentManagerTest):
         # default replies hide flagged comment
         self.assertEqual(self.parent_comment_2.replies().count(), 2)
         self.assertEqual(self.parent_comment_2.replies(include_flagged=True).count(), 3)
-
-    def test_reaction_signal(self):
-        """Test reaction model instance is created when a comment is created"""
-        parent_comment = self.create_comment(self.content_object_1)
-        self.assertIsNotNone(Reaction.objects.get(comment=parent_comment))
-        # 1 reaction instance is created for every comment
-        self.assertEqual(Reaction.objects.count(), self.increment)
-
-    def test_flag_signal(self):
-        """Test flag model instance is created when a comment is created"""
-        current_count = Flag.objects.count()
-        parent_comment = self.create_comment(self.content_object_1)
-        self.assertIsNotNone(Flag.objects.get(comment=parent_comment))
-        # 1 flag instance is created for every comment
-        self.assertEqual(Flag.objects.count(), current_count + 1)
 
     @patch('comment.models.comments.hasattr')
     def test_is_flagged_property(self, mocked_hasattr):
@@ -260,24 +245,6 @@ class ReactionInstanceModelTest(BaseCommentManagerTest):
         self.create_reaction_instance(self.user, self.comment, self.LIKE)
         self.assertRaises(IntegrityError, self.create_reaction_instance, self.user, self.comment, self.DISLIKE)
 
-    def test_post_save_signal_increases_count_on_creation(self):
-        """Test reaction count is increased on creation"""
-        comment = self.comment
-        self.create_reaction_instance(self.user, self.comment, self.LIKE)
-        comment.refresh_from_db()
-        self.assertEqual(comment.likes, 1)
-        self.assertEqual(comment.dislikes, 0)
-
-    def test_post_delete_signal_decreases_count(self):
-        """Test reaction count is decreased when an instance is deleted"""
-        comment = self.comment
-        instance = self.create_reaction_instance(self.user, self.comment, self.LIKE)
-        comment.refresh_from_db()
-        self.assertEqual(comment.likes, 1)
-        instance.delete()
-        comment.refresh_from_db()
-        self.assertEqual(comment.likes, 0)
-
     def test_comment_property_likes_increase_and_decrease(self):
         """Test decrease and increase on likes property with subsequent request."""
         comment = self.child_comment_2
@@ -366,22 +333,6 @@ class ReactionModelTest(BaseCommentTest):
 
         self.assertEqual(self.comment_1.reaction.dislikes, 0)
 
-    def test_increase_reaction_signal(self):
-        self.assertEqual(self.comment_1.reaction.likes, 0)
-        # reaction instance created
-        reaction_instance = ReactionInstance.objects.create(
-            reaction=self.comment_1.reaction, user=self.user_1, reaction_type=1)
-        self.comment_1.reaction.refresh_from_db()
-        self.assertEqual(self.comment_1.reaction.likes, 1)
-        self.assertEqual(self.comment_1.reaction.dislikes, 0)
-
-        # edit reaction instance won't change reaction count
-        reaction_instance.reaction_type = 2  # dislike
-        reaction_instance.save()
-        self.comment_1.reaction.refresh_from_db()
-        self.assertEqual(self.comment_1.reaction.likes, 1)
-        self.assertEqual(self.comment_1.reaction.dislikes, 0)
-
 
 class ReactionInstanceManagerTest(BaseCommentTest):
     def test_clean_reaction_type(self):
@@ -406,28 +357,8 @@ class FlagInstanceModelTest(BaseCommentFlagTest):
         comment.refresh_from_db()
         self.assertEqual(comment.flag.count, 1)
 
-    def test_post_delete_signal_decreases_count(self):
-        """Test flag count is decreased when an instance is deleted"""
-        data = self.flag_data
-        comment = self.comment
-        instance = self.create_flag_instance(self.user, comment, **data)
-        comment.refresh_from_db()
-        self.assertEqual(comment.flag.count, 1)
-        instance.delete()
-        comment.refresh_from_db()
-        self.assertEqual(comment.flag.count, 0)
-
-    def test_post_save_signal_increases_count_on_creation(self):
-        comment = self.comment
-        self.create_flag_instance(self.user, comment)
-        comment.refresh_from_db()
-        self.assertEqual(comment.flag.count, 1)
-
 
 class FlagInstanceManagerTest(BaseCommentFlagTest):
-    def setUp(self):
-        super().setUp()
-
     def test_clean_reason_for_invalid_value(self):
         data = self.flag_data
         data.update({'reason': -1})
@@ -494,20 +425,6 @@ class FlagModelTest(BaseCommentFlagTest):
     def test_comment_author(self):
         comment = self.comment
         self.assertEqual(comment.user, comment.flag.comment_author)
-
-    def test_increase_flag_signal(self):
-        self.assertEqual(self.comment.flag.count, 0)
-        # instance created
-        self.set_flag(self.user, self.comment, **self.flag_data)
-        self.comment.flag.refresh_from_db()
-        self.assertEqual(self.comment.flag.count, 1)
-        # instance edited won't increase the flag count
-        flag_instance = FlagInstance.objects.get(user=self.user, flag__comment=self.comment)
-        self.assertIsNotNone(flag_instance)
-        flag_instance.info = 'change value for test'
-        flag_instance.save()
-        self.comment.flag.refresh_from_db()
-        self.assertEqual(self.comment.flag.count, 1)
 
     def test_is_flagged_enabled(self):
         flag = self.create_comment(self.content_object_1).flag
